refactor: 이미지 객체 내부 인스턴스 변수 갯수 줄이기 위한 리팩토링
